% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev.R
\name{bgev}
\alias{bgev}
\alias{dbgev}
\alias{pbgev}
\alias{qbgev}
\alias{rbgev}
\title{The blended generalised extreme value distribution}
\usage{
dbgev(
  x,
  location,
  scale,
  shape,
  pa = 0.05,
  pb = 0.2,
  alpha = 0.5,
  beta = 0.5,
  log = FALSE
)

pbgev(
  q,
  location,
  scale,
  shape,
  pa = 0.05,
  pb = 0.2,
  alpha = 0.5,
  beta = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

qbgev(
  p,
  location,
  scale,
  shape,
  pa = 0.05,
  pb = 0.2,
  alpha = 0.5,
  beta = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

rbgev(n, location, scale, shape, pa = 0.05, pb = 0.2, alpha = 0.5, beta = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location, scale, shape}{location, scale and shape parameters.}

\item{pa, pb, alpha, beta}{Gumbel to GEV mixing parameters; see Details.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
\eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dbgev} gives the density,
\code{pbgev} gives the distribution function,
\code{qbgev} gives the quantile function, and
\code{rbgev} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for 
the blended generalised extreme value distribution with parameters 
\code{location}, \code{scale}, \code{shape}, \code{pa}, \code{pb}, 
\code{alpha} and \code{beta}.
}
\details{
The blended generalised extreme value distribution with \code{location} 
parameter \eqn{q_\alpha}, \code{scale} parameter \eqn{s_\beta} and 
\code{shape} parameter \eqn{\xi} has cumulative distribution function
given by
\deqn{H(x \mid q_\alpha, s_\beta, \xi) = F(x \mid q_\alpha, s_\beta, \xi)^{p(x')} G(x \mid \tilde q_\alpha \tilde s_\beta)^{1 - p(x')}}
where 
\deqn{F(x \mid q_\alpha, s_\beta, \xi) = \exp \left\{ - \left[ \dfrac{x - q_\alpha}{s_\beta(\ell_{1 - \beta / 2, \xi} - \ell_{\beta / 2, \xi})^{-1}} + \ell_{\alpha, \xi} \right]^{-1/\xi}_+  \right\}}
with \eqn{\ell_{a, \xi} = (-\log a)^{-\xi}}, \eqn{[x]_+ = \max(0, x)},
\code{alpha} and \code{beta} parameters \eqn{\alpha} and \eqn{\beta}, 
respectively, 
\deqn{G(x \mid q_{\tilde \alpha}, s_{\tilde \beta}) = \exp\left\{ -\exp\left[- \left\{\dfrac{x - \tilde q_\alpha}{\tilde s_\beta(\ell_{1 - \beta / 2} - \ell_{\beta / 2})^{-1}} + \ell_{\alpha} \right\}\right]\right\}}
with \eqn{\ell_a = \log(-\log a)},
\deqn{\tilde q_\alpha = a - \dfrac{(b - a)(\ell_\alpha - \ell_{p_a}))}{\ell_{p_a} - \ell_{p_b}}~~\text{and}~~\tilde s_\beta = \dfrac{(b - a)(\ell_{\beta / 2} - \ell_{1 - \beta/2})}{\ell_{p_a} - \ell_{p_b}},}
with \eqn{a = F^{-1}(p_a \mid q_\alpha, s_\beta, \xi)}, 
\eqn{b = F^{-1}(p_b \mid q_\alpha, s_\beta, \xi)}, with \code{pa} and \code{pb}
parameters \eqn{p_a} and \eqn{p_b}, respectively, and where 
\eqn{x' = (x - a) / (b - a)} and \eqn{p(x)} denotes the cumulative distribution 
of the Beta(5, 5) distribution.

Default values for \code{pa}, \code{pb}, \code{alpha} and \code{beta}
are taken from Castro-Camilo et al. (2022).
}
\examples{

dbgev(3, 2, 1, .1)

}
\references{
Castro-Camilo, D., Huser, R., & Rue, H. (2022). Practical strategies for 
generalized extreme value-based regression models for extremes. 
Environmetrics, 33(6), e2742. \doi{10.1002/env.2742}
}
\seealso{
\link{predict.evgam}
}
