% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgev.R
\name{bgev2gev}
\alias{bgev2gev}
\alias{gev2bgev}
\title{Conversion between blended generalised extreme value distribution and 
generalised extreme value distribution parameters}
\usage{
bgev2gev(
  location,
  scale,
  shape,
  pa = 0.05,
  pb = 0.2,
  alpha = 0.5,
  beta = 0.5,
  simplify = FALSE
)

gev2bgev(
  location,
  scale,
  shape,
  pa = 0.05,
  pb = 0.2,
  alpha = 0.5,
  beta = 0.5,
  simplify = FALSE
)
}
\arguments{
\item{location, scale, shape}{location, scale and shape parameters.}

\item{pa, pb, alpha, beta}{Gumbel to GEV mixing parameters; see Details.}

\item{simplify}{logical; should \code{simplify2array()} be called at the end?}
}
\value{
A \code{list} or \code{array} if \code{simplify = TRUE}.
}
\description{
Conversion between location, scale and shape parameters of the generalised
extreme value distribution and corresponding parameters of blended 
generalised extreme value distribution in both directions.
}
\details{
The blended generalised extreme value distribution has \code{location} 
parameter \eqn{q_\alpha}, \code{scale} \eqn{s_\beta} and 
\code{shape} \eqn{\xi} parameters; see, e.g., \code{dbgev}. The generalised 
extreme value distribution's location and scale parameters in its typical 
form are given by \eqn{\mu = q_\alpha - s_\beta (\ell_{\alpha, \xi} - 1) / (\ell_{1 - \beta/2, \xi} - \ell_{\beta/2, \xi})}
and \eqn{\sigma = \xi s_\beta / (\ell_{1 - \beta/2, \xi} - \ell_{\beta/2, \xi})}. 
So \code{bgev2gev} gives the mapping \eqn{(q_\alpha, s_\beta, \xi) \mapsto (\mu, \sigma, \xi))}.
The reverse mapping, \eqn{(\mu, \sigma, \xi) \mapsto (q_\alpha, s_\beta, \xi))},
is given by \code{gev2bgev}.
}
\examples{

bgev2gev(2, 1, .1)
gev2bgev(2, 1, .1)

}
\references{
Castro-Camilo, D., Huser, R., & Rue, H. (2022). Practical strategies for 
generalized extreme value-based regression models for extremes. 
Environmetrics, 33(6), e2742. \doi{10.1002/env.2742}
}
\seealso{
\link{dbgev}
}
