% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{condexd0}
\alias{condexd0}
\alias{condexd12}
\alias{condexd34}
\title{Conditional extreme value model negative log-likelihood}
\usage{
condexd0(pars, X1, X2, X3, X4, ymat, xmat, dupid, dcate, nhere)

condexd12(pars, X1, X2, X3, X4, ymat, xmat, dupid, dcate, nhere)

condexd34(pars, X1, X2, X3, X4, ymat, xmat, dupid, dcate, nhere)
}
\arguments{
\item{pars}{a list of vectors of coefficients for each bGEV parameter}

\item{X1}{a design matrix for (transformed) alpha}

\item{X2}{a design matrix for (transformed) beta}

\item{X3}{a design matrix for mu}

\item{X4}{a design matrix for (transformed) sigma}

\item{ymat}{a matrix}

\item{xmat}{a matrix}

\item{dupid}{a scalar or vector, identifying duplicates in Xs; -1 corresponds to no duplicates}
}
\value{
condexd0 a scalar, the negative log-likelihood

condexd12 a matrix, first then second derivatives w.r.t. parameters

condexd34 a matrix, third then fourth derivatives w.r.t. parameters
}
\description{
Conditional extreme value model negative log-likelihood
}
\examples{
## to follow
}
