% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evgam-package.R
\docType{package}
\name{evgam-package}
\alias{evgam-package}
\title{evgam: A package for fitting generalised additive extreme value models.}
\description{
The evgam package's main function is evgam.
}
\section{evgam functions}{

The evgam functions ...
}

\author{
\strong{Maintainer}: Ben Youngman \email{b.youngman@exeter.ac.uk} (\href{https://orcid.org/0000-0003-0215-8189}{ORCID})

Authors:
\itemize{
  \item Ben Youngman \email{b.youngman@exeter.ac.uk} (\href{https://orcid.org/0000-0003-0215-8189}{ORCID})
}

}
\keyword{internal}
