% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FCtmax}
\alias{FCtmax}
\title{Fort Collins, Colorado, US daily max. temperatures}
\format{
A data frame with 18156 rows and 2 variables

The variables are as follows:

\describe{
  \item{date}{date of observation}
  \item{tmax}{daily maximum temperature in degrees Celcius}
}
}
\usage{
data(FCtmax)
}
\description{
Daily maximum temperatures at Fort Collins, Colorado, US 
from 1st January 1970 to 31st December 2019
}
\examples{

library(evgam)
data(FCtmax)
}
\keyword{datasets}
