% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{fitted.evgam}
\alias{fitted.evgam}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{evgam}(object, ...)
}
\arguments{
\item{object}{a fitted \code{evgam} object}

\item{...}{not used}
}
\value{
Fitted values extracted from the object `object'.
}
\description{
Extract Model Fitted Values
}
\examples{

data(fremantle)
fmla_gev <- list(SeaLevel ~ s(Year, k=5, bs="cr"), ~ 1, ~ 1)
m_gev <- evgam(fmla_gev, fremantle, family = "gev")
fitted(m_gev)

}
