% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoevgam.R
\name{plot.geoevgam}
\alias{plot.geoevgam}
\title{Plot results of a geoevgam fit}
\usage{
\method{plot}{geoevgam}(
  x,
  type = c("threshold", "excess"),
  view = c("polar", "cartesian"),
  norm,
  margins = "original",
  ...
)
}
\arguments{
\item{x}{An object of class \code{geoevgam}.}

\item{type}{Character vector specifying plot types: "threshold", "excess", "gauge", or "qqplot".}

\item{view}{Character specifying the coordinate system for gauge plots: "polar" or "cartesian".}

\item{norm}{Numeric; the norm to use for the gauge plot. Defaults to the norm used in \code{object}.}

\item{margins}{Character; margin type for QQ-plots. Defaults to "original".}

\item{...}{Additional arguments passed to the underlying plot functions.}
}
\description{
Plot results of a geoevgam fit
}
\examples{

n <- 1e3
gauss_sim_data <- rmvnorm(n, numeric(2), matrix(c(1, .8, .8, 1), 2))
laplace_sim_data <- qlaplace(pnorm(gauss_sim_data))
polar1 <- polarise(t(laplace_sim_data), norm = 2)
data1 <- data.frame(radius = polar1$r, angle = polar1$phi[1, ])
fit1 <- geoevgam(data = data1)
plot(fit1)
plot(fit1, type = 'gauge', col = 2, lwd = 2, view = 'cartesian')
par(mfrow = c(2, 2))
plot(fit1, type = 'qqplot')
plot(fit1, type = 'qqplot2')
plot(fit1, type = 'qqplot', margins = 'uniform')
plot(fit1, type = 'qqplot2', margins = 'uniform')

\donttest{
# example using L_1 norm
polar2 <- polarise(t(laplace_sim_data), norm = 1)
data2 <- data.frame(radius = polar2$r, angle = polar2$phi[1, ])
fit2 <- geoevgam(data = data2, norm = 1) 
# passing norm ensures plotted gauge is correct
plot(fit2, type = 'gauge', col = 2, lwd = 2, view = 'cartesian')
}

}
