% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoevgam.R
\name{polarise}
\alias{polarise}
\title{Convert Cartesian coordinates to Polar coordinates}
\usage{
polarise(x, norm = 2)
}
\arguments{
\item{x}{A matrix with exactly 2 rows representing the x and y coordinates.}

\item{norm}{A numeric value specifying the L-p norm for calculating the radius. Defaults to 2 (Euclidean).}
}
\value{
A list containing:
\item{r}{The calculated radii (distances from origin).}
\item{phi}{The angles in radians, adjusted to the range [0, 2*pi].}
}
\description{
Convert Cartesian coordinates to Polar coordinates
}
