% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/exi.R
\name{runmax}
\alias{runmax}
\alias{dfrunmax}
\title{Running maximum}
\usage{
runmax(y, n)

dfrunmax(data, cons, ynm, n = 2)
}
\arguments{
\item{y}{a vector}

\item{n}{an integer giving the number of observations to calculate running maxmimum over; defaults to 2}

\item{data}{a data frame}

\item{cons}{a character string for the variable in \code{data} that identifies consecutive observations}

\item{ynm}{a character string for the variable in \code{data} that is the observations}
}
\value{
\code{runmax} returns a vector of the same dimension as \code{y}

\code{dfrunmax} returns a data frame with observations swapped for \eqn{n}-observation running maximum
}
\description{
Running \eqn{n}-value maximum and data frame with variable swapped for running maximum
}
\examples{
runmax(runif(10), 5)

}
