% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{seq_between}
\alias{seq_between}
\title{More Sequence Generation}
\usage{
seq_between(x, length = NULL)
}
\arguments{
\item{x}{a 2-vector}

\item{length}{an integer}
}
\value{
A vector
}
\description{
Generate a sequence of values between a range.
}
\examples{

seq_between(c(1, 9))
seq_between(range(runif(10)), 5)

}
\seealso{
\link[base]{seq}, \link[base]{seq_len}, \link[base]{seq_along}
}
