% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{laplace}
\alias{laplace}
\alias{dlaplace}
\alias{plaplace}
\alias{qlaplace}
\alias{rlaplace}
\title{The Laplace Distribution}
\arguments{
\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{location}{location parameter (mean).}

\item{scale}{scale parameter (must be positive).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], 
otherwise, P[X > x].}
}
\description{
Density, distribution function, quantile function and random generation for 
the Laplace distribution with location \code{location} and scale \code{scale}.
}
