% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{Laplace}
\alias{Laplace}
\alias{dlaplace}
\alias{plaplace}
\alias{qlaplace}
\alias{rlaplace}
\title{The Laplace Distribution}
\usage{
dlaplace(x, location = 0, scale = 1, log = FALSE)

plaplace(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qlaplace(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rlaplace(n, location = 0, scale = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location}{location parameter (mean).}

\item{scale}{scale parameter (must be positive).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], 
otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random generation for 
the Laplace distribution with location \code{location} and scale \code{scale}.
}
