% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.evgam}
\alias{print.evgam}
\title{Print a fitted \code{evgam} object}
\usage{
\method{print}{evgam}(x, ...)
}
\arguments{
\item{x}{a fitted \code{evgam} object}

\item{...}{not used}
}
\value{
The call of the \code{evgam} object
}
\description{
Print a fitted \code{evgam} object
}
\examples{

data(fremantle)
fmla_gev <- list(SeaLevel ~ s(Year, k=5, bs="cr"), ~ 1, ~ 1)
m_gev <- evgam(fmla_gev, fremantle, family = "gev")
print(m_gev)

}
